# assets/openapi.yaml
openapi: 3.1.0
info:
  title: openccu-loom REST API
  version: 1.0.0
  description: |
    North-bound REST + WebSocket surface of the openccu-loom daemon.
    Errors follow RFC 9457 problem+json.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
  - openIdConnect: [viewer]
...
paths:
...
  /devices/{addr}:
    get:
      tags: [devices]
      summary: Single device detail
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDevice
      summary: Remove a device from the registry (admin)
      tags: [devices]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "202":
          description: Unpair dispatched asynchronously
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    patch:
      operationId: patchDevice
      summary: Update mutable device metadata (name, room assignment)
      tags: [devices]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [operator]
      parameters:
        - $ref: "#/components/parameters/Address"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        "202":
          description: Update accepted
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /admin/values-cache/stats:
    get:
      operationId: getValuesCacheStats
      summary: Persistent VALUES-cache statistics (admin-only)
      description: |
        Snapshot of the wire-DP VALUES cache: row count, JSON byte
        size, and the cumulative restore/cast/gc/flush counters since
        process start. See ADR 0018.
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      responses:
        "200":
          description: Current cache statistics.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValuesCacheStats" }
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: cache feature disabled or store unwired.
...
  /admin/values-cache/reset:
    post:
      operationId: resetValuesCacheGlobal
      summary: Drop every row from the persistent VALUES cache (admin-only)
      description: |
        Removes ALL cached wire-DP values. The next cold boot starts
        with `source=unobserved` on every data point until live
        events repopulate them.
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      responses:
        "204":
          description: cache cleared.
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: cache feature disabled or store unwired.
  /devices/{addr}/values-cache/reset:
    post:
      operationId: resetValuesCacheDevice
      summary: Drop the VALUES-cache rows for one device (admin-only)
      tags: [devices]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      parameters:
        - in: path
          name: addr
          required: true
          schema: { type: string }
      responses:
        "204":
          description: device rows cleared.
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: device address unknown.
        "500": { $ref: "#/components/responses/InternalError" }
        "503":
          description: cache feature disabled or store unwired.
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Username + password (operator-grade or above for mutations).
        The role hierarchy (viewer ⊂ operator ⊂ admin) is enforced
        server-side regardless of authentication scheme.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token issued via the auth tokens surface (admin-only
        provisioning). Token's stored role determines authorisation.
    openIdConnect:
      type: openIdConnect
      description: |
        OpenID Connect authentication. The operator configures the
        OIDC discovery URL; the placeholder below is the standard
        well-known path. Per-operation `security:` blocks express
        the **minimum role** the daemon enforces via OIDC-style
        scopes — the same enforcement applies when the client
        authenticates via basicAuth or bearerAuth, but those schemes
        cannot natively express scopes in OpenAPI 3.x.

        Scope vocabulary mirrors `Identity.role`:
          - `viewer` — read-only operations
          - `operator` — mutations (write data points, manage links,
            edit schedules, run config-import, trigger refresh)
          - `admin` — dangerous operations (delete device, edit log
            levels, list users/tokens, trigger backup, manage
            interfaces, edit Matter fabrics)
      openIdConnectUrl: /.well-known/openid-configuration
  parameters:
    Address:
      name: addr
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: Validation or parsing failure
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    BadGateway:
      description: Upstream error (CCU unavailable, write failed)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ServiceUnavailable:
      description: Backend not wired
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    InternalError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    UnprocessableEntity:
      description: Semantic validation failure
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 9457 problem detail. The `type` URI selects one of the
        canonical problem kinds the daemon emits; the `code` member
        carries the short tag (also surfaced as the `X-Problem-Code`
        response header). Clients should switch on `code` for typed
        error handling — the `type` URI is the human-readable form.

        **Rate-limiting:** when the daemon emits `code: rate_limited`
        (HTTP 429) it sets the `Retry-After` response header per
        RFC 9110 §10.2.3 with the integer number of seconds the
        client should wait before retrying. Clients that respect
        the header are immune to the throttle; clients that don't
        will see repeated 429s with the same `Retry-After` value
        until the throttle window expires.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI for one of the canonical problem kinds. All
            URIs share the base `https://openccu-loom.dev/errors/`
            and the trailing segment matches the `code` enum below.
            The base URI is a stable identifier, not a fetchable
            resource — it is not guaranteed to resolve to a document.
          enum:
            - https://openccu-loom.dev/errors/validation
            - https://openccu-loom.dev/errors/not_found
            - https://openccu-loom.dev/errors/conflict
            - https://openccu-loom.dev/errors/unauthorized
            - https://openccu-loom.dev/errors/forbidden
            - https://openccu-loom.dev/errors/unsupported
            - https://openccu-loom.dev/errors/rate_limited
            - https://openccu-loom.dev/errors/internal
            - https://openccu-loom.dev/errors/bad_request
            - https://openccu-loom.dev/errors/service_unready
            - https://openccu-loom.dev/errors/upstream_unavailable
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Short tag (also surfaced as the `X-Problem-Code` response header).
          enum:
            - validation
            - not_found
            - conflict
            - unauthorized
            - forbidden
            - unsupported
            - rate_limited
            - internal
            - bad_request
            - service_unready
            - upstream_unavailable
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
              min: {}
              max: {}
...
    DeviceSummary:
      type: object
      required: [address, interface, model, name, available, channels_count]
      properties:
        address: { type: string }
        interface: { type: string }
        interface_id: { type: string }
        model: { type: string }
        sub_model: { type: string }
        name: { type: string }
        manufacturer: { type: string }
        product_group: { type: string }
        available: { type: boolean }
        channels_count: { type: integer }
        updatable: { type: boolean }
        rooms:
          type: array
          items: { type: string }
...
    DeviceDetail:
      allOf:
        - $ref: "#/components/schemas/DeviceSummary"
        - type: object
          properties:
            firmware:
              type: object
              properties:
                Current: { type: string }
                Available: { type: string }
                Updatable: { type: boolean }
                UpdateState: { type: string }
            availability:
              type: object
              properties:
                IsReachable: { type: boolean }
                LastUpdated: { type: string, format: date-time }
                BatteryLevel: { type: integer, nullable: true }
                LowBattery: { type: boolean, nullable: true }
                SignalStrength: { type: integer, nullable: true }
            channels:
              type: array
              items: { $ref: "#/components/schemas/ChannelSummary" }
    ChannelSummary:
      type: object
      required: [address, number, paramset_key, data_points_count]
      properties:
        address: { type: string }
        number: { type: integer }
        paramset_key: { type: string }
        data_points_count: { type: integer }
        custom_dp_name:
          type: string
          description: |
            Stable name of the Custom-DP this channel is attached to.
            Empty when the channel is not owned by any CDP. Lets the
            SPA's CDP-first Übersicht view (ADR 0016) filter channels
            that are already represented by a CDP tile.
...
    ValuesCacheStats:
      type: object
      required: [rows, value_json_bytes, restored_rows, cast_failures, gc_rows_deleted, flush_batches, flushed_entries]
      properties:
        rows: { type: integer, format: int64, description: "Current row count in the cache." }
        value_json_bytes: { type: integer, format: int64, description: "Sum of value_json byte sizes across all rows." }
        restored_rows: { type: integer, format: int64, description: "Cumulative rows successfully restored since process start." }
        cast_failures: { type: integer, format: int64, description: "Cumulative entries whose cached value could not be coerced into the data point's static type." }
        gc_rows_deleted: { type: integer, format: int64, description: "Cumulative rows removed by the dead-row GC since process start." }
        flush_batches: { type: integer, format: int64, description: "Cumulative SaveBatch transactions committed since process start." }
        flushed_entries: { type: integer, format: int64, description: "Cumulative entries written across all SaveBatch transactions." }
...